beerocks_agent&controller: amxrt destruction order
Amxrt destructor call amxb_be_remove_connections to free all connections.
AmbiorixImpl/AmbiorixConnection destructor call amxb_free to free one entry.
Amxrt require to destruct after AmbiorixImpl/AmbiorixConnection.

Due to the use of shared_ptr, it's not straightforward to determine which
destruct first. In order to encounter scenario like below example:
    shared_ptr db;
    shared_ptr Amxrt;
    shared_ptr AmbiorixImpl;
    db->hold_ptr(AmbiorixImpl)

The only exception is the global variable in WHM/bpl.cpp, it depends on
the linker that whether the `guarantee` or the `m_ambiorix_cl` construct
first.

This change uses a ptr to hold reference count, to ensure the Amxrt must
destruct after others in the same compiling scope;
Register a `atexit` hook to implicitly destruct `m_ambiorix_cl`, which release
the reference count it hold to the connection.

closes: PPM-3032

Signed-off-by: Lu Dai <lu.dai@mind.be>