beerocks_controller: sigmngr list circled
the `amxd_dm_init(&rt.dm)` is not re-enterable by it's implementation.
it's downcall `amxp_sigmngr_init(&dm->sigmngr)` is using a list append
method which is not robust enough, will create the circled list like
the example:
  (gdb) p *(amxc_llist_it_t*)0x7f96a9b2a8
  $5 = {next = 0x7f96a9b2a8 <rt+416>, prev = 0x7f96a9b2a8 <rt+416>,
    llist = 0x7f96acc168 <amxp_sigctrl+16>}

iterating this list will cause a endless loop.

this patch avoid the duplicate calls of `amxd_dm_init`, considering
that AmbiorixImpl is contructed after Amxrt.

Signed-off-by: Lu Dai <lu.dai@mind.be>