slave: fill actions upon m8 reception
Current situation is that M8 handling triggers placeholders functions in agent.

This commit adds handling for enabling disabling EP (the function in sta_wlan_hal_whm is to be develloped).
It also adds handling for bsta_reconfiguration from ap_autoconfiguration to backhaul manager. The backhaul manager handling comes in a future commit.

Signed-off-by: Florent Couzon <prplmesh.opensources@sagemcom.com>