Easymesh Plus: Agent Operating Class Reporting
Currently, there is no control to check for unavailable
overlapping channels in operating classes that use central
channels, which leads to incorrect operating class reporting.

To resolve this, the proposed method involves checking whether
there are any unavailable overlapping channels. If such a
channel is detected, the channel that belongs to the operating
class that uses the central channel should be added to the
non-operable channels list.

Example: if channel 144 is set to non-operable for operating
class 121 (20MHz) and 123 (40MHz), channel 138 should be set as
non-operable as well due to utilizing the same spectrum with
channel 144.

PPW-389

Signed-off-by: Arda Aras <arda.aras@airties.com>