bcl: use SOCK_CLOEXEC flag when create socket
socket file descriptor should better be closed on exec, if it does not
serve the purpose as inter process communication channel in between
parent process and child process. one of the benefits is to better manage
the resource-indexed by file descriptor, release it as expected.

this fix appends the `sock_cloexec` flag when create socket.

Signed-off-by: Lu Dai <lu.dai@mind.be>