experiment: use SOCK_CLOEXEC flag when create tcp socket
in beerocks, fronthaul process, as the child, clone fds from agent process,
as the parent. which leaves tcp socket created by agent not release after
it exits, leads to tcp connection fail, obseved in UCC service.

MR3833 set the `cloexec` flag by `fcntl`, but not robust enough, there is
a chance that `fork` called between `socket` and `fcntl`

this fix set the `sock_cloexec` flag when create the socket

Signed-off-by: Lu Dai <lu.dai@mind.be>