ci: merge train fails on haze
this is a fix without concrete code analysis!

CI is failing the merge train on HAZE platform, by bisect we located a cluster of commits.
one suspicious change is the `multi band operation`. it is disabled by default and was disabled,
but enabled after one recent change. by observation, failure pattern as:
  1, for one commit, haze-MAP-4.3.1 sometimes work, sometimes fail
  2, sometimes it report no `topology response` sniffered, somtimes no `ap autoconfiguration`

MBO attribute is used by pWHM and then synchronize to hostapd configuration. Choose to disable this
attribute based on several reasons, even with out concrete proof:
  1, we didn't enable this attribute before
  2, no requirement asks to support MBO
  3, by experiment, disable it and then test case works as expected

Signed-off-by: Lu Dai <lu.dai@mind.be>