agent: crash on tlvf_airties_utils destruct
using current libamxb, AmbiorixConnection and Amxrt has competition to free amxb_bus_ctx_t.
airties is using global static objects in header, which has no way to guarantee the desturct
order. core dump files can be found on jira ticket PPM-3218.

this fix choose a minimal footprint approach to fix the issue, using reference to avoid the on-exit
destruction of the airties objects.

note that the airties `global static objects in header` really needs to be improved, refer
to coverity report with filter `impact=high, type=Initialization or destruction ordering is unspecified`.
this fix didn't touch that issue but left a comment for others information.

closes: PPM-3218

Signed-off-by: Lu Dai <lu.dai@mind.be>