Agent: EasyMesh Plus: Channel Selection Interface Implementation between prplAgent and pwhm
This commit provides an interface between the agent and PWHM to
forward the content of channel selection request messages.

It parses the channel selection request message based on rank
values of zero, then creates an acs_list that includes the operating
class, excluded channels, and the length of excluded channels.

The acs_list is then sent from the backhaul manager to the slave
and subsequently to the ap manager.

Example format:

{
  "acs_list": [
    {
      "opclass": 81,
      "exclude_channels_length": 5,
      "exclude_channels": [
        1,
        2,
        3,
        4,
        5
      ]
    },
    {
      "opclass": 83,
      "exclude_channels_length": 2,
      "exclude_channels": [
        1,
        2
      ]
    }
  ]
}

PPM-3254

Signed-off-by: Arda Aras <arda.aras@airties.com>