link_metrics_collection: Explicitly send the beacon request with the values, received with beacon_metrics_query
Explicitly send another beacon request with the values packed in
beacon_metrics_query. With the introduction of beacon reports multi-channel support we send only requests
for the channels that are packed inside chan_report_list, but not for
the channel inside the beacon metrics query. The 4.7.9 test expects the
beacon request with the beacon metrics query.

PPM-3318

Signed-off-by: Denys Stolbov <d.stolbov@inango-systems.com>