CI: bump prplOS version to 5 June 2025
Bumps the prplOS version used in CI for regular and fast track platforms to f578eef4.
This fixes an issue where builds in prplMesh CI are failing because a checksum mismatch. (PPW-570)

Also upsteps easymesh_cert and the OSP builder version to allow nighly CI runs on OSPv1 and OSPv2.

Signed-off-by: Maarten De Decker <maarten.dedecker@mind.be>