controller: add_rsn_parameters_configuration_tlv : add new RSN tlv
WPA3 Personal Compatibility Mode security is not yet supported.
To support it we need to add new RSN parameters configuration tlv.

If an agent supports RSN overriding, controller must
send rsn parameters configuration tlv.

In this commit, we only use this new tlv for
WPA3 Personal Compatibility mode configuration.

We keep the old way for other security modes.

For this new tlv, security data needs to be filled.
We predefined this data in structure.

For agent supporting RSN Overriding, we also add the BSS index
which is the new way to identify BSS through several TLVs.

https://prplfoundationcloud.atlassian.net/browse/PPM-3290

Signed-off-by: Vincent Ayme <vincent.ayme@sagemcom.com>