bwl: whm: send channel preference report to controller
As per the Easymesh Standard, if a Multi-AP-Agent detects a change in DFS
state of a channel it should send the unsolicited Channel Preference report
to the Controller.

As per the Airties requirements, whenever the Agent detects radar it should
send Channel Preference Report message indicating that radar has been found
both in Channel Preference TLV and CAC Status TLV, which has been done
earlier for RDK Platform using the CSA_Finished event.

use the existing flow for this requirement to send the report to Controller
whenever CSA_Finished event with reason "RADAR" is received.

use RadarTriggeredDfsChannels parameter in ChannelMgt object to get the Radar
triggered channels for filling Channel Preference and CAC Status TLV.

Closes PPM-3278

Signed-off-by: rramasamy <rramasamy@maxlinear.com>