agent: channel_selection_task: fix reason code updating
add reason code - operation disallowed due to radar detected for higher
frequency operating classes like 128 and 129, if any 20 MHz is radar detected.

previously it uses the last 20MHz channel preference key, so that even if
the first 20MHz is affected by radar detection and the last 20MHz is radar
free it adds the higher preference.

Closes PPM-3375

Signed-off-by: rramasamy <rramasamy@maxlinear.com>