vendor_message_slave: VendorMessageSlave(): fix uninitialized bridge MAC.
The vendor daemon was not functioning correctly because db->bridge.mac
was not initialized.

The root cause is that the AgentDB is not shared between the
beerocks_agent and beerocks_vendor_message processes. As a result,
initializing db->bridge.mac in the agent process does not reflect in
the vendor daemon.

To fix this: Reading the Mac address from the DM 'X_PRPLWARE-COM_Agent.Info.MACAddress'

This ensures proper filtering of incoming messages and correct handling
in the vendor-specific message logic.

Legacy Build: Added a dummy files for build-in-docker-legacy was failing.

closes: PPM-3394
Signed-off-by: Sundram Patel <Sundram.p@tataelxsi.co.in>
(cherry picked from commit 47af4c6e77cbcb600b706182b161ac51ea5f5bb2)
Signed-off-by: Cihangir Odabas <cihangir.odabas@prplfoundation.org>