agent: tlvf_airties_utils: Incorrect device_role value in Device Info TLV
A new management_mode option has been added to the UCI config to identify
non-prplMesh controllers. For instance, when a non-prplMesh controller is
in use, the mode should be set to "Non-Prpl-Controller-and-Agent" to
indicate that the device is managed by an external, non-prplMesh platform.
This would ensure controller role would be picked up by local_gw.

Closes PPM-3396

Signed-off-by: Sangeetha <sangeetha.s1@tataelxsi.co.in>