controller: topology_task: discard false neighbors
Sometimes MAC addresses belonging to a 1905 device are reported as non1905 neighbors
by some Agents. Therefore some Agents 1905 are added in the database as neighbors.

One solution could be to prevent Agent from sending false neighbor to Controller, but
Agents don't have the list of all the Agents. The other solution is to have the
Controller filter the list of neighbors upon reception.

We choose to filter the list of neighbors in the Controller in handle_topology_response.

https://prplfoundationcloud.atlassian.net/browse/PPM-3456
Signed-off-by: Florent COUZON <florent.couzon-ext@sagemcom.com>