agent: arp sent with no IP
When a client is connected but does not have an IP address, the monitor sends an
ARP request for the broadcast IP address. Since stations do not respond to such
requests, there is no point in sending them; instead, the monitor should log
the event and not send ARPs in this case.

Closes: PPM-2505

Signed-off-by: Dmytro Kysil <d.kysil@inango-systems.com>