controller: db: set wifi channel for non Intel Agt
Currently, the radio wifi channel in radio database is set at
reception of AP-Autoconfiguration with Intel Slave join, when a
DFS CAC is performed or when a CSA happens. When a non Intel
Slave joins the radio wifi channel stays empty in database.

We could add the set in non Intel Slave join, but the channel
information is not available at AP-Autoconfiguration reception.
Or we can change it upon reception of the operating channel
report following the onboarding.

Add the set of radio wifi channel while setting the
op classes.

Closes: PPM-3459

Signed-off-by: Florent Couzon <florent.couzon-ext@sagemcom.com>