controller: op class: add BH op class
tr 181 implements a CurrentOperatingClassProfile in
DataElements.Network.Device.MultiAPDevice.Backhaul.
This is currently not implemented in the code.

Create DM node and fuse the filling of
DataElements...Radio.CurrentOperatingClasses and
Backhaul.CurrentOperatingClassProfile.
In addition the operating class profiles are stored in
database.

In order to avoid unecessary DM nodes creations and destructions,
create all DM nodes instances (for Radio and Backhaul) the first
time we receive an Operating Channel Report. There is one node
for each possible bandwith (currently 5). The removal of the nodes
is now a clear of the database and DM values.

Closes: PPM-3466

Signed-off-by: Florent Couzon <florent.couzon-ext@sagemcom.com>

sqash