tlvf_airties_utils: cache device&ethernet stats TLVs to speed up AP Metrics gen
Problem:
When creating the AP Metrics Response, the parameters contained in the
vendor specific TLVs like Device Metrics and Ethernet Stats TLVs are
read from the data model. Reading from the data model is relatively slow,
so generating the AP Metrics Response can be delayed by about 200-300 ms.
This delay makes a big difference especially when the AP Metrics Response
period is as short as 1000 ms.

Solution:
A cache mechanism has been added to tlvf_airties_utils. This mechanism
pre-caches the parameters read from the data model that are included
in the Device Metrics and Ethernet Stats TLVs. When generating the
AP Metrics Response, using the cached data to create the 1905 message
significantly speeds up the process. According to measurements,
generating the vendor-specific TLVs now takes about 4 ms, which means
at least a 50x speed improvement compared to the previous approach.

This commit also includes a fix for the mismatch between the radio UID in
the Airties Device Metrics TLV and the UID used by the prplmesh agent.

PPM-3393
PPM-3473

Signed-off-by: Durmus Koyuncu <durmus.koyuncu@airties.com>