Draft:Controller:Update parent/child details of Agent with Wired Connection
Currently we have parsed the backhaul details of agent that is connected
via wireless backhaul from Topology response. However there is no direct
mechanism to derive backhaul details of an agent that is connected via
Wired backhaul.

Added logic to parse the interface lists, neighbor information and iterate
all information on database to derive the hierarchy of wired connected agents.

Fixes: PPM-2043
Signed-off-by: Anguru Suresh <suresh.anguru@tinno.us>