controller: database: bSTA flag
The bSta flag is used to tag if a Sta is BH or not. This flag is
set to true for BH Sta during onbording but it is then set
to false by handle_topology_notification. This behavior,
at minimum, prevents BH steering from working.

The actual condition to set the flag to false is:
bSta && !sta_in_handoff (sta beeing steered). This part of the
code has been integrated 6 years ago (commit: 57a1007) without
further explanations. On top of in the current code, no bSta can
have the sta_in_handoff set to true, meaning any bSta will
automatically be tagged as non bSta.

Remove this part that makes no sense in the current
implementation.

Closes: PPM-3462

Signed-off-by: Florent COUZON <florent.couzon-ext@sagemcom.com>