agent: implement reason codes to indicate DFS state in preference report
Problem: When the controller sends a Channel Selection Request to the agent to switch
to a DFS channel, the agent reports the target channel in the Operating Channel Report
even though the current channel has not yet changed. This leads to inaccurate reporting
during the CAC period.

Possible Solution: According to section 17.2.13 (Channel Preference TLV) of EasyMesh
specification, the reason code can be used to indicate the DFS channel state.

 - 1001: Immediate operation possible on a DFS channel (CAC has
 been run and is still valid and channel has been cleared for use)

 - 1010: DFS channel state unknown (CAC has not run or its validity
 period has expired)

This addition allows the controller to factor the DFS state into its decision-making.

Closes: PPW-517

Signed-off-by: Arda Aras <arda.aras@airties.com>