ap_wlan_hal_whm: qos: get APs paths by RadioReference
Previously QOS_MAP_SET was set for ALL APs. It's wrong.
The logic is refactored to set map only to specific
radio APs.

Right now pWHM has RadioReference paths saved without "Device."
So, Ref values in prplMesh have to be stripped as well to match
pWHM side. This is temporary workaround and will be handled
in scope of PPM-3533 in the future

Closes PPM-3530

Signed-off-by: Volodymyr Pavlenko <pavlenko.vv@pm.me>