bwl: whm: set timeout for 11k request to 0
Problem:
If prplMesh or higher layer doesn't set a timeout in the arguments and
triggers sendRemoteMeasumentRequest, the default timeout is set to 1000ms in
pwhm. This causes a delay in receiving the Beacon report in prplMesh.

Fix:
Set the default timeout to 0ms to eliminate unnecessary delay and ensure
faster Beacon report retrieval.

Closes : PPW-818

Signed-off-by: rramasamy <rramasamy@maxlinear.com>