CI: platform: do not match regex rules for nightly jobs
Gitlab started matching regex rules using the !~ operator, in variables that are not defined in a pipeline.
This has the consequence that merge train piplines launch nightly jobs, without defining a test list or platform.
Merge train pipelines fail because these unwanted nightly jobs are failing.

Fix this by changing from a regex rule to a check if the nighly variables exist.

Closes: PPM-3541

Signed-off-by: Maarten De Decker <maarten.dedecker@mind.be>