link_metrics_collection: Add WFA exception to multi-chan 11k Bcn Req handling
Previous [CORE_0036] implemented the 1905-spec interpretation:
- When Channel is 255     - send measurement requests only for channels listed in
  the Channel Report List.
- When Channel is not 255 - send a single measurement request for the specified channel.

A later commit(see PPM 3318) changed this behaviour to also send a request for
Channel 255 itself in addition to the listed channels, to satisfy WFA certification tests.

This patch restores the original 1905-spec behaviour and adds a special case
for WFA certification tests, allowing an extra request for Channel 255 only when
running in certification test mode.

PPM-3548

Signed-off-by: Durmus Koyuncu <durmus.koyuncu@airties.com>