Agent: EasyMesh Plus: Predictable & Persistent SSID Provisioning
This commit allows the creation of a persistent and predictable link
between BSSID interface on a platform and EM controller SSID profile.

The mechanism is primarily achieved by mapping the type metadata parameter
contained in the WSC attribute to the Device.WiFi.SSID.*.CustomAlias platform parameter.

If agent receives an Airties BSS Type WSC attribute, it configures the corresponding
BSS and then retrieves the VAP id according to the BSS type.

Closes: PPM-3464

Signed-off-by: Arda Aras <arda.aras@airties.com>
Signed-off-by: Volodymyr Pavlenko <pavlenko.vv@pm.me>