controller: autoconfiguration: add default SecMode for legacy agents
EM R6.1 introduces a new flag in the AP Capability TLV,
Support for RSN Overriding
An Agent sets the flag to true to advertise support for
RSN Overriding.
If the Agent does not support the feature,
or uses an old definition of the TLV,
in both cases the flag is 0.

In case the Controller has only RSN Security Configurations,
and is onboarding an Agent that does not support them,
push to this agent the security mode from the legacy
RSN IE: WPA2 for 2.4/5GHz, or WPA3 for 6GHz

Implements: PPM-3352

Signed-off-by: Iacob Juc <iacob.juc_ext@softathome.com>