controller : controller database wifi_channel
For non Intel Slaves wifi_channel is added to radio with a primary
channel and the highest bandwith. Currently the wifi_channel set
in handle_current_op_class is primary channel and always 20MHz.

Update the logic in handle_current_op_class to set the wifi_channel
to match the one already in place.

Closes : PPM-3466
Signed-off-by: Florent Couzon <florent.couzon-ext@sagemcom.com>