agent: backhaul_manager: reset correctly when the profile changes
In R6 certification, the testbed will trigger multiple certification_mode resets, to different Multi-AP profiles (mapr3, mapr6).
In case the agent is still in reset state when the second reset is triggered, this profile change is currently discarded.

Solution: check if the Multi-AP profile has changed, compared to last reset.

Fixes: PPM-3627

Signed-off-by: Maarten De Decker <maarten.dedecker@mind.be>