controller/nbapi: fix get_param_uint32() return type
The function was likely supposed to return a
uint32_t but it was returning bool. It was
handling uint32_t internall, but it inadvertently
ended up casting it to bool.

This means the function would always ever return a
0 or 1 for any callers trying to use it as
uint32_t. The compiler didn't complain making this
a silent problem.

This seems to have affected Spatial Reuse
requests. Must've slipped in testing.

Found while trying to re-use it for
ChannelSelectionRequest implementation.

Signed-off-by: Michal Kazior <michal@plume.com>