controller: fix the channel selection request tlv construction
Technically multiple Channel Preference TLV
entries can be within a message. These could
technically be for the same radio. Prplmesh Agent
currently has arguably non-compliant behavior to
EM spec, and clears out preferences for a given
radio whenever it processes (next) Channel
Preference TLV.

This changes makes the Channel Preference TLV to
be geerated _once_, and have op_classes stuffed to
it iteratively instead of having multiple
TLVs with 1 op class each.

Even if there are Agents that can (properly)
adhere to the spec and not clear preference on
each TLV processing for a given radio, it's safer
to generate the message this (new) way. It just
has better guarantees to work even for
semi-compliant Agents.

Signed-off-by: Michal Kazior <michal@plume.com>