controller: rework ChannelSelectionRequest to ingest htable
This finally matches the input arrangement that
TR-181 is expecting of ChannelSelectionRequest.

The only problem is this is not properly enforced
data model schema. It's a free-form json-like
argument: htable.

This is fine for a stop-gap for now. Final calls -
when proper schema representation is implemented -
will take compatible input, or near identical.

An example way of calling this is (channel 6,
20MHz, preference=14):

> ba-cli 'Device.WiFi.DataElements.Network.Device.1.Radio.1.ChannelSelectionRequest(Class={1={OpClass=81,Channel={1={Channel=6,Preference=14}}}})'

Signed-off-by: Michal Kazior <michal@plume.com>