beerocks/scripts: add example script to control channel of an agent
This script is intended to serve as a reference or
an example of how ChannelSelectionRequest may be
used to force an EasyMesh Agent to a specific
channel.

The script employs a strategy where it marks all
non-target channels as preference=0, the main
target channel as preference=14, and optionally
beacon channel as preference=7.

Example use of the script against a 2.4GHz Radio
(indice 1) of some Device (indice 1), to tell it
to go to channel 6 at 40MHz is:

> /opt/prplmesh/scripts/prplmesh_set_chan.sh 1 1 83-6

Signed-off-by: Michal Kazior <michal@plume.com>