agent: handle Backhaul Steering Request changing radios
This is incomplete.

The get_wireless_hal() return a null pointer for
non-main bSTA interface. This is because there's a
code which drops non-selected one.

Keeping them at all times may not work, although
it probably can be made to work. This is what
needs to be looked at and fixed. Once that is done
I suspect the logic in this patch will just work.

Signed-off-by: Michal Kazior <michal@plume.com>