controller: implement ChannelSelectionRequest() for data model
The USP TR-181 2.19 introduced an async command to
relay channel preferences to other mesh agents via
DataElements. This maps to EasyMesh Channel
Selection Request message along with Channel
Preference TLV.

Immediate use case is to instruct an agent to move
its radio(s) to another channel, restrict it from
using some channels, or prefer a particular subset
of them. This can be used to drive optimizations
externally over data model.

This clashes with controller's DCS which may
override Channel Preference TLVs. To avoid that
the operator or the integrator is expected to
disable DCS. For example, this can be done with:

> ba-cli 'X_PRPLWARE-COM_Controller.Configuration.DynamicChannelSelectionTaskEnabled=0'

Due to Ambiorix and ODL limitations the data model
schema uses htable for function arguments.

Example of relaying preference=5 for channel=6 on
2.4GHz at 20MHz is:

> ba-cli 'Device.WiFi.DataElements.Network.Device.1.Radio.1.ChannelSelectionRequest(Class={1={OpClass=81,Channel={1={Channel=6,Preference=5}}}})'

References: PPM-3650
Co-authored-by: Shivam Thakur <sthakur@plume.com>
Signed-off-by: Shivam Thakur <sthakur@plume.com>
Signed-off-by: Michal Kazior <michal@plume.com>