agent: signal scan failures and retry
There are rare cases where the scan completes
without a success. These cases are often transient
and retrying the scan will yield a success.

The ScanResults event is tied to Backhaul Steering
Request state machine. If the ScanResults is not
delivered the entire state machine stops and the
only remaining outcome is a timeout.

Instead of that try to retry the scan hoping the
next scan will yield a ScanResults.

This makes Backhaul Steering Request handling more
robust and predictable.

References: PPM-3661
Signed-off-by: Michal Kazior <michal@plume.com>