bwl: whm: respect connect() channel
If connect()-ing across channels worked for anyone
it must've worked on accident / due to implicit
HAL driver behaviour that is able to drag the
local APs to the (new) target channel.

This isn't guaranteed to make connect() succeed
immediately, or at all. It does increase the
likeliness of it working.

Ideally this should involve a CSA (Channel Switch
Announcement) on the local APs to reduce impact on
the fronthual and downstream backhaul traffic.
There's no direct CSA control in the DM so the
best that can be done is to rely on
Device.WiFi.Radio.{i}.Channel and hope the stack
can make the thing thing.

References: PPM-3661
Signed-off-by: Michal Kazior <michal@plume.com>