bwl: agent: fix amxb_usp events not received
Problem:
--------

The agent process is not receiving pwhm events over usp direct socket.

Cause:
------

beerocks_agent connects simultaneously to both ubus and usp direct
socket.
The following initialization sequence exposes several limitations in
libamxb and amxb_usp:

1 - Load and connect to ubus
2 - Load usp backend
3 - Set usp configuration
4 - Connect to usp socket

In step 1:

An internal libamxb configuration is created without a usp section.

In step 2:

- When loading usp backend the previous configuration (which doesn't
include any usp section) is automatically set to amxb_usp. As a
"limitation" in amxb_usp, this will result into a NULL amxb_usp
configuration (ie config_opts = NULL).

- As a consequence no EndPointID (needed later by usp LocalAgent to
created imtp handshake) can be created and later events subscription
will fail internally. The fail is not returned by the subscribe API
which is yet another problem.

In step 3:

- Calling amxb_set_config has no effect because the configuration is
applied only to backends listed in the load_order.

- Since the usp backend was not added to load_order list (internal to
libamxb) in step 2, the USP configuration is never applied. (This
appears to be another limitation in libamxb.)

Solution:
---------

Invert steps 2 and 3.

By setting the usp configuration before loading the backend, the usp
section is present in the internal libamxb configuration. As a result,
when the usp backend is loaded, it receives a non-NULL configuration,
allowing the EndPointID to be created correctly and enabling successful
event subscription.

Fixes: PPW-1208
Signed-off-by: Houssem Dafdouf <houssem.dafdouf_ext@softathome.com>