agent: use newly exposed bwl helpers for security_type
Instead of duplicating the mapping use a shared
one. The AgentDB is already making use of WiFiSec
so it is already coupled with bwl component. No
sense in trying to abstract _another_ string
converter if one can be shared.

This will allow saving WPS / EasyMesh provision
credentials into the DM that will be done in
follow up commits.

References: PPM-3661
Signed-off-by: Michal Kazior <michal@plume.com>