agent: backhaul: handle credential update better
When WSC message is received during runtime it was
possible to spawn extra EndPoint Profile(s) on
other Radios, possibly in a broken state.

The actul intention behind this message is really
to store the credentials and possibly re-evaluate
backhaul Profile, if necessary.

This relies implicitly on the other change that
adds hooks into the DM to listen for updates of
Agent.Configuration and to re-apply them to the
backhaul.

References: PPM-3661
Signed-off-by: Michal Kazior <michal@plume.com>