bwl: whm: add freq_type fallback for get_scan_results()
Turns out the underlying code tries to use
CentreChannel as a center_frequency argument to
construct WifiChannel and infer freq_type out of
it. This (obviously) fails with freq_type getting
set to FREQ_UNKNOWN accompanied by a set of
warnings in logs like this:

> ERROR 15:31:38:259 <548543425184> beerocks_wifi_channel.cpp[298] --> Center frequency (6) of channel 6 is below the minimum
> ERROR 15:31:38:259 <548543425184> beerocks_wifi_channel.cpp[23] --> invalid wifiChannel params. Creating an empty channel instead
> ERROR 15:31:38:259 <548543425184> beerocks_wifi_channel.cpp[298] --> Center frequency (6) of channel 6 is below the minimum
> ERROR 15:31:38:259 <548543425184> beerocks_wifi_channel.cpp[23] --> invalid wifiChannel params. Creating an empty channel instead
> ERROR 15:31:38:259 <548543425184> beerocks_wifi_channel.cpp[298] --> Center frequency (6) of channel 6 is below the minimum

Ideally pWHM would need to expose center frequncy
(MHz) or band in scan results.

In the meantime infer this best-effort based on
radio band type.

This prevented select_bssid() from working because
it saw _zero_ AP candidates it could connect to.

References: PPM-3661
Signed-off-by: Michal Kazior <michal@plume.com>