agent: backhaul: try recovering from lost EndPoint Profiles
If EndPoint Profiles (for whatever reason) got
lost, and agent was (re)started (eg. after a
reboot) the system would remain in a stuck state,
waiting for WPS.

This is actually something that could happen if
Profile processing was caught mid-way by a service
restarts (which can technically be done through
DM, remotely by the operator). This could leave
devices strangled / stuck in a disconnected state
until user re-pairs the extender.

This needs to distinguish between provisioned and
non-provisioned device. This is why the default
SSID in Agent configuration is now empty. As per
the other commits this is now properly stored
after WPS completes.

Signed-off-by: Michal Kazior <michal@plume.com>