agent: fix back_radio ssid/pass initiation
Contrary to what it might seem like, the 2nd
argument for std::string constructor of this
arrangement is *not* a strnlen() limit. It's an
actual "this string is N bytes of this buffer"
*regardless* if there are \0 bytes there or not.

In other words, these std::string constructors
were always setting up N byte long buffers. This
meant that even if the underlying c-style
null-terminated string were 0-length, the
std::string was always N-long.

This meant .empty() was never true. Comparing to
"" was always false.

This doesn't seem to be diractly causing problems,
yet. I want to rely on SSID presence in
back_radio to infer if extender was provisioned
prior (eg. WPS) and try to do select_bssid() to
find best candidate on start instead of getting
back into WAIT_WPS.

References: PPM-3661
Signed-off-by: Michal Kazior <michal@plume.com>