agent: backhaul: skip non-matching ssids
There's absolutely no need to try connecting to
BSSes that have non-matching SSIDs.

What this can actually even result in, due to pWHM
and/or wpa_supplicant behaviour, is a perpetual
scan loop that prevents doing scans from the Agent
causing it to get stuck with a bad BSSID forever
until it is re-provisined from scratch, eg. with
WPS.

Signed-off-by: Michal Kazior <michal@plume.com>