controller: fix agent monitoring setup on restart
When controller is restarted (which can happen
remotely, as per operator request) it would fail
to properly initiate some functions.

For example Backhaul Sta Capability would remain
unknown because it was never asked for. This in
turn caused some AgentDB parts to be left
unpopulated, like backhaul_station_mac.

This in turn would cause all sorts of weird
problems. For example SteerWiFiBackhaul() needs to
infer what the bSTA MAC is in order to construct
the 1905 CMDU. Without that, it can't construct
it.

Without this change unless Agents are also
restarted, and/or everything is restarted in
tandem, the system (the controller notably) would
be stuck in a weird initialized-but-incomplete
state.

This also makes edit-compile-test development
cycles more reliable.

References: PPM-3661
Signed-off-by: Michal Kazior <michal@plume.com>